# docs/05-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Klick-Pro (servio) HTTP API
  version: 0.0.0-reverse-engineered
  description: >-
    Reverse-engineered from source at commit cd8f4fb (verified 2026-09-16). Source of truth is the implementation; every
    operation carries x-source file:line.


    Conventions and caveats:

    - Responses are raw NextResponse.json; errors are mostly {"error": string}. The apiSuccess/apiError envelope in
    src/lib/api-response.ts is not used.

    - Every POST/PUT/PATCH/DELETE under /api/* must carry an Origin header equal to the app origin or APP_URL, otherwise
    proxy.ts returns 403 {"error":"Request origin is not allowed."}. This applies to webhooks and non-browser clients
    too. Under the custom server only the exact APP_URL origin passes in practice; a genuine same-origin request to a
    different host (e.g. 127.0.0.1 when APP_URL is localhost) also gets 403. [VALIDATED 2026-09-17 · V-20]

    - /api/v1/:path* is rewritten to /api/:path* (next.config.ts, afterFiles). Physical routes /api/v1/messages and
    /api/v1/professionals take precedence and have no /api/* twin. /api/v1/v1/* is not rewritten twice and returns
    404. [VALIDATED 2026-09-17 · V-13]

    - Schemas marked [NEEDS VALIDATION] could not be fully established from code; no properties were invented.

    - Realtime Socket.IO events (path /api/realtime) are documented in docs/05-api/api-specification.md, not here.


    Human-readable reference: docs/05-api/api-specification.md
servers:
...
paths:
...
  /api/search:
    get:
      tags:
        - Public
      operationId: quickSearchJobs
      summary: Quick search of open jobs (max 8)
      x-source: app/api/search/route.ts:4-27
      security: []
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Jobs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobs
                properties:
                  jobs:
                    type: array
                    maxItems: 8
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type:
                            - string
                            - "null"
                        category:
                          type:
                            - string
                            - "null"
                        locationLabel:
                          type:
                            - string
                            - "null"
...
  /api/v1/messages:
    get:
      tags:
        - Messaging
      summary: Contact list with conversations, or a single conversation when conversationId is given
      operationId: getMessages
      security:
        - cookieAuth: []
      description: >-
        Source app/api/v1/messages/route.ts:27. Physical v1 route (no /api/messages). ADMIN sees all users and
        conversations.
      parameters:
        - name: conversationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Conversation (with conversationId) or contacts
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      conversation:
                        allOf:
                          - $ref: "#/components/schemas/SocketConversation"
                          - type: object
                            properties:
                              messages:
                                type: array
                                maxItems: 100
                                items:
                                  $ref: "#/components/schemas/SocketMessage"
                  - $ref: "#/components/schemas/MessageContactsResponse"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Conversation access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - Messaging
      summary: Mark messages read (one conversation or all) and emit message:read
      operationId: markMessagesRead
      security:
        - cookieAuth: []
      description: Source app/api/v1/messages/route.ts:209. No zod; invalid JSON gives 500.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: string
                  description: Required unless all=true
                all:
                  type: boolean
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
        "400":
          description: Conversation is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Conversation access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Messaging
      summary: Send a direct message (non-admin pairs need a running project)
      operationId: sendMessage
      security:
        - cookieAuth: []
      description: >-
        Source app/api/v1/messages/route.ts:275. Creates or reuses a SocketConversation and creates a SocketMessage.
        Sends NEW_MESSAGE notification and email, and message:new to both parties. No length limit, no rate limit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipientId
                - text
              properties:
                recipientId:
                  type: integer
                text:
                  type: string
                  description: Trimmed, non-empty; no max length
                job:
                  type: string
                  description: Conversation label override
                projectId:
                  type: integer
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: string
                  message:
                    $ref: "#/components/schemas/SocketMessage"
        "400":
          description: Recipient and message are required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Messaging is available for running projects only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Recipient is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: servio_session
      description: >-
        HS256 JWT {userId, role, sessionId} issued by src/lib/auth.ts createSession; validated against the sessions
        table on every request. Same cookie for clients, professionals and admins — role is inside the token.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Same session JWT sent as Authorization: Bearer. Accepted only by the few handlers noted per operation."
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    SocketConversation:
      type: object
      description: "Prisma SocketConversation row [NEEDS VALIDATION: full columns]"
      properties:
        id:
          type: string
          format: uuid
        userAId:
          type: integer
        userBId:
          type: integer
        userAName:
          type: string
        userBName:
          type: string
        userAAvatarUrl:
          type:
            - string
            - "null"
        userBAvatarUrl:
          type:
            - string
            - "null"
        job:
          type: string
        updatedAt:
          type: string
          format: date-time
    SocketMessage:
      type: object
      description: "Prisma SocketMessage row [NEEDS VALIDATION: full columns]"
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        senderId:
          type: integer
        receiverId:
          type: integer
        body:
          type: string
        readAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
    ProjectSummary:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        title:
          type: string
        category:
          type:
            - string
            - "null"
        status:
          type: string
        progress:
          type: integer
        completedMilestones:
          type: integer
        totalMilestones:
          type: integer
        isCompleted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    MessageContactsResponse:
      type: object
      properties:
        role:
          type: string
          enum:
            - CLIENT
            - PROFESSIONAL
            - ADMIN
        contacts:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              firstName:
                type: string
              lastName:
                type: string
              avatarUrl:
                type:
                  - string
                  - "null"
              role:
                type: string
              name:
                type: string
              conversationId:
                type:
                  - string
                  - "null"
              lastMessage:
                oneOf:
                  - type: "null"
                  - $ref: "#/components/schemas/SocketMessage"
              unreadCount:
                type: integer
              projects:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
              activeProject:
                oneOf:
                  - type: "null"
                  - $ref: "#/components/schemas/ProjectSummary"
...